langchainAgent/langchain_class/requirements.txt
langchain_community
langchain_google_genai 
python-dotenv
langchain_core 
pydantic
langchain_huggingface 
sentence-transformers